Create a Browser class and global instance
This refactors the code in 'browser.js' into a class, in preparation
of adding more complex logic around browser interactions.  At the
expense of invoking the wrath of the gods, I also create a global
instance of the Browser class.  The way the class is defined, it
automatically communicates with the embedded <browser> in our XUL; if
new Browser objects are created they will deal with this same browser.
So instead of having to write 'new Browser()' all over the place, I
think it is more straight-forward to simply have a global for the
Browser.

Note, however, that other Browser objects *could* point to different
<browser> elements by changing the 'widget' property.  Not sure why
would ever do so, though.