Create a class for representing and convert CSS color values
This introduces a Color class with three main properties: red, green,
and blue, which are integers for representing a color value.  The
class is geared towards converting back and forth between the various
color representations allowed by CSS.

Right now only colors in the format 'rgb(...)' are supported.  While
testing out ideas I discovered that the 'CSSStyleDeclaration.color'
property of the DOM always represents these colors in 'rgb(...)'
format---or at least that was true of the sites I tested.

The class comes with a generic from_css() constructor that should be
used in favor of creating the objects via 'new', eg.:

  var color = Color.from_css(
      document.styleSheets[0].cssRules[0].style.color
  );