Create makefile for deployment
This file will download and compile everything we need,
simple go to the deployment directory and run:
make all

It will then download the source for the nullsoft installer and
compile it. It also downloads the xulrunner files which need to be
included in the deployment packages. Finally it puts together the packages
themselves in the distros folder.

Currently it only builds the tarball for linux builds since we haven't
written the installer scripts for the nullsoft installer for windows builds.