Gets ColorProperty to record the original colorFormat the color came in
Each ColorProperty instance has a 1:1 relationship with a color property
declaration in the CSS. The original declaration had a certain string
format. Now ColorProperty records that format.

This functionality does NOT belong in the Color class, because an instance
of Color has a 1:N relationship with original delclarations in the CSS.

This commit also refactors the getColorString and setColor methods so that
the switch statement is called once, on construction, rather than every
time getColorString or setColor is called.