Refactor use of hex patterns in Color class
First, we eliminate the hexPattern regex.  We already have
shortHexPattern and longHexPattern, representing the '#fff' and
'#ffffff' formats' respectively.  If we need to know whether a string
is one or the other then we can match against both.  Combining the two
by hand is error-prone.

Second, we introduce an isHexFormat() function to replace a match
against the removed hexPattern.

Finally, we change read_hex() to use the short and long patterns
instead of duplicating their contents.