Refactor Palette.insertStyle() into a loop
For each of the various properties 'color', 'backgroundColor', etc.,
we need to call insertProperty() if 'style' has that property.
Originally this was done by manually checking

  if ( style.backgroundColor ) { ... }

and so on for each property.  However, we can take advantage of the
fact that in Javascript

  object['p']

is the same thing as

  object.p

allowing us to create an array of property names that we loop over.