Do not create ColorProperty getters and setters with a switch-case
Previously the ColorProperty constructor would take a second argument
that would be an enumerated value for color properties, like
'backgroundColor'.  The constructor would switch on this value to
determine how to define setColor() and getColorString().  By passing
in a string naming the property instead we can eliminate the need for
the switch-case and remove a large chunk of duplicated code.

This requires a change to the call-points where ColorProperty objects
are created.  Fortunately this only happens in the Palette class, so
the change here is easily accomodated.