Add 'File' menu with a 'Save' option that writes "Hello world!"
This patch introduces the CSApplication.getSaveFile() method for
opening a save file dialog box that we can use for selecting CSS files
for output.  The dialog initially restricts the user to only selecting
CSS files, but they can alternatively choose to save to any file, per
the usual UI for their platform.  The method accomplishes this by
going through the nsIFilePicker interface and returns the resulting
nsILocalFile---if possible.  It does not tell us whether the user is
saving a new file or over-writing an existing one.

This also implements the very basic CSApplication.save() method for
what will eventually write out our modified CSS files.  Currently it
does nothing but call getSaveFile() and then writes out "Hello world!"
to the output.  No action is taken when the user cancels the
operation.

A menu bar is created along the top of the application, with one
entry: 'File'.  Which itself only has one option: 'Save'.  A new
<commandset> is added to store all of the commands that will be bound
to menu options.  Currently that set contains only one command, which
binds the 'Save' option to CSApplication.save().

Redmine-Ticket: 1516