Allow Ctrl+S as a short-cut for saving
This commit makes it possible to use the standard Ctrl+S short-cut for
bringing up the save dialog.  In addition, 's' is now an access key on
the File menu, so that a user my click File and then press 's' to
save.

Note that in order for this to work, it is important that the <keyset>
is defined _before_ the <menubar>.  This is because the <menuitem> for
Save refers to <key id="save-key"> in order to setup the short-cut
that makes "Ctrl+S" appear on the menu.  So for that to work we need
the key short-cut to be defined first.

Redmine-Ticket: 1516