[1516] Add "Save" button to each stylesheet, to save formatted output
This adds a save button to each sheet, which uses an ugly oncommand
handler to handle each button press.

This also manages to add pretty formatting to each stylesheet that
is saved, in the form:

body{
    color: #FFFFFF;
    font-family: sans;
}

p{
    color: #000000;
}

Each stylesheet is also given a name, which in most cases is the name
of the original file on the server, or "style_tag_5.css" for sheets that
come from inline style tags. The original full name of the sheet (href)
is shown in a tooltip.

This is going to need to change to accommodate the refactored stylesheet
class when we get it.