Allow users to quit the application via standard practices
Many applications with a 'File' menu add a 'Quit' option at the very
end of it that will close everything.  This patch adds such an option,
which calls the new CSApplication.quit() method.

It also binds the Ctrl+Q sequence to quit the app as well,
e.g. Firefox and Thunderbird.