Teach Elements to remove all of their child nodes
When we call updateCSSFilePane() we grab the <vbox> element that holds
all off the CSS files and delete the child nodes one by one.  It seems
likely enough that in the future we will have similar lists that we
will want to clear out and then update.  To facilitate, the loop from
updateCSSFilePane() is removed and the Element prototype is extended
with a function for removing all child elements.