[1537] Wrap various DOM related objects
This commit encapsulates a lot of changes throughout the system and is
not completely finished. The primary emphasis on this change is to abstract
style rules so that other objects no longer directly access the DOM. This
also allows implementation of a swatch-level undo/redo system which is currently
partially completed. Please review the actual changes for a more thorough
explanation