Get Colors to return their color names, if applicable.
Note - some colors, such as "cyan" and "aqua" are synonymous. This
method does not differentiate between such colors, and naively returns
the last match.

Nevermind the fact that this doesn't shortcut the loop on the first
match - someone please find a more efficient way to match the
values than looping.