Hook in new and improved undo and redo into the UI
Get each hbox that contains a set of swatch controls - the parent
element for the set of related UI widgets for a swatch - responsible
for updating its own contents.

Make use of the "relatedObjects" array to hook up the element's
children, so it can update its component parts. This moves UI
functionality from the event handler into the GUI element.

The swatch UI element has two update methods - one for the input
textbox, and one for everything else. Sometimes we want to update
everything but the textbox, so we don't trigger another update,
creating a cycle of updates, or locking the undo chain.

The event handler updateSwatchWithExplicitValue is retained, as
this is the event handler for reading from the text box. We may
later have handlers for updating from tools, as opposed to updating
with an explicit text value.

Update the undoSwatch handler, and add the redoSwatch handler, which
_should_ update the textbox.

Teach Swatch to better track its colorFormat, and to take an optional
initialization parameter "colorFormat".

Teach Swatch to properly undo and redo, based on Chris Rasys' work.
Also create the distinction between swatch.setColor(color) and
swatch.updateColor(color). The first sets a history state. The second
simply updates the current state and attached DOM objects without setting
a history state. This will be useful for slider controls, where we
don't want history states for every intermediate value on the slider.