Refactor interface for stylesheets
This patch is prep-work for extending our interface around
stylesheets.  The first thing I did was go in and delete everything I
could without breaking the existing functionality of the application.
In particular:

  * Deleted the 'csStylesheetRule' class.

  * Deleted a number of private members from all classes.

  * Renamed a lot of 'originalFoo' to 'foo'.  This was for
   constructors that save a copy of their argument.

  * Renamed 'csStylesheet' to 'csStyleSheet' for consistency.  The
    word is written as 'StyleSheet' almost every where else, including
    in official DOM classes.

  * Deleted csStyleSheet.getResultingSourceCode().  It was not being
    used as far as I could tell.  Although, realistically, it will
    come back.

Then I documented the interfaces to the classes.  While ripping things
out I triggered some errors in both csDocument and in the Palette
classes.  I had to teach csDocument to:

    1. Make sure 'sheet' exists before referencing 'sheet.cssRules'.

    2. Call insertStyle() with just 'rule' instead of 'rule.style'.

The problem in the Palette class was similar to the first error in
csDocument, just had to make sure something was there before using
it.  Originally that problem was triggered because the public
interface to csStyleRule had changed and calls to property() were
failing, but I corrected that---and that being the case, the change to
Palette can possibly be removed.

The file now come close to passing JSLint.

One other functional change is that now importedSheets is itself an
array of csStyleSheet objects, instead of CSSStyleSheet objects.

Redmine-Ticket: 1206
Redmine-Ticket: 1208
Redmine-Ticket: 1537