Make application title bar include website title
Whenever we browse to another site, we change the app title to
"CSSchemer - Foo Site", or whatever the site name is.  Initially I
thought this could be done by changing the title inside of
onLocationChange(), but then we have a strange value on startup.  But
it turns out onStateChange() works just fine.

Redmine-Ticket: 1561