Add keyboard short-cuts for Forward and Back navigation
This adds support for the usual Firefox short-cuts for moving forward
and back and the browser history.  But while I was at it, I also added
support for my own prefered keys (analogous to Emacs).  Therefore,
this patch provides these keys for these actions:

    * Go Back
        - Alt + Left Arrow
        - M-b

    * Go Forward
        - Alt + Right Arrow
        - M-f

I _guess_ we could get rid of the second form.

Redmine-Ticket: 1563