Teach csStyleSheet to fetch unmodified CSS text
This relies primarily on the Ajax method, which was originally
Richard's idea.  Once we have a valid 'href', we are able to grab the
source as text/plain content via an XML HTTP requests.  Although not
in the real spirit of Ajax, this fetch is performed synchronously.

There is a chance that the request may return non-CSS content and we
have to test for that.  So the String prototype is extended with a
method for testing whether or not it represents a valid CSS
stylesheet.

Redmine-Ticket: 1206