Teach csDocument.LoadStyleSheets how to handle nested csStyleSheets
Commit a2cd64d0 ('Refactor interface for stylesheets') changed the
structure of csStyleSheet.importedSheets to be an array of
csStyleSheet objects.  Previously it was an array of sheet names, and
the LoadStyleSheets function would create csStyleSheet objects for
those names as it looped over them, putting them in the sheetsToAdd
queue.  However, since csStyleSheets already performs that logic,
creating the appropriate objects, we can put the importedSheets
directly into the list of instantiated sheet objects and by-pass that
need for sheetsToAdd.