Refactor Swatch as Broadcaster of its own state
This cleans up the public interface of the swatch class, demarcating
public and private members.  It also clears up some poorly defined
functionality.

It adds the ability for the swatch class to broadcast two kinds of
state changes to registered listeners.

One kind of state change is
a transient color change to the liveEditColor. This is to be used
for editing the swatch when no history state is desired, such as
while adjusting a slider in the GUI.
Observers can register to be notified of such changes.

The other kind of change is the history change. This changes the actual
recorded state of the swatch, adding to or modifying the history and
current color.

Observers can register to be notified of either one or both of these
changes.