Add swatch selection handling
This gets each Swatch to track it's own selection state, and broadcast
changes to that state to any registered observer.

The palette is registered as an observer, and as selections are made,
the palette tracks those changes. The palette can also make groups
selections and deselections.

The UI is also registered as an observer of the selection state.
Specifically, the XBL binding for the Swatch Control is registered
as an observer, and updates its appearance based on the selection
state of the swatch.

The XBL swatch controls also handle events, making selections based
on mouse input. So far, regular click and accel-click (ctrl-click)
events are handled in a standard fashion.