Teach isValidStylesheet() to remove comment blocks first
The initial version of isValidStylesheet() had an error in the logic
that tried to search through comments while looking for '<'
characters.  If we see that character in a comment then we can't say
the stylesheet is invalid.  But instead of trying to keep track of
that, a more simple solution is to remove all comments before looking
for the '<' character.