Add an XBL color picker, refactoring the XBL bindings for modularity
This is inspired by the ColorJack color picker, and uses its
base images. The code is totally different, but the underlying
principle of how it functions is the same.

This also refactors the XBL bindings to allow the kind of modularity
used in the color picker. It moves the XBL to its own directory,
rather than having it in the root directory. There is a subdirectory
for the independent "ColorPickerPlus" binding, which does not
contain dependencies on the rest of the project.

See the following site for ColorJack
http://www.colorjack.com/software/dhtml+color+picker.html