Handle mouseover selected swatches with an animation
This adds an observer system for swatch mouseover events.
It is used here for highlighting the whole selection at once.
The observer system could also be used later to change
other aspects of the UI depending on what swatch is hovered over.

The palette animates when the user mouses over a selected swatch.
The animation is all at the paletteControl level, and only affects
swatchControls because of the CSS rules in place.

This method of animation may be glitchy, and as of yet, the most
recently selected swatchControl isn't added to the selected
swatches styles, for some reason.