Add base observablePicker class for internal picker communication
The different kinds of pickers can all broadcast in the same way,
broadcasting their drag and select events.

A ColorPickerPlus is composed of two sub-pickers - a two dimensional
picker and a vertical picker. Each of them extends the
observablePicker to get their drag and select notification
functionality. In addition, the ColorPickerPlus as a whole is an
extension of "observablePicker", itself broadcasting its drag and
select changes.