Begin textbox RGB and HSV inputs. HSV works. RGB doesn't.
This is not efficient, because if you hold down an arrow key on
an input, each change is registered as a new color selection, which
will all get added to the undo states of the swatch. We need to get
a keypress held down to register as a drag somehow. Maybe it only
registers a selection when the user unfocuses the text input area.