Get RGB input working.
This gets the RGB input fields working. Like the previous commit,
this creates a new swatch state for every keyboard (or mouse) change
to the text field. When someone presses the up arrow, it creates a
ton of swatch history states. These should be drag states, like the
previous commit message said. This will become obvious once the
undo functionality is added back in.