Get clustering to work.
This finishes the clustering algorithm. It is a complete linkage clustering
algorithm using HCL distance as the distance metric. We may want to expirement
with single linkage or UPGMA style clustering to see if it yields better
results.

Tuning the cluster distance threshold makes a difference in the cluster
size, as does tuning the gamma parameter of the HCL color space.