Use a percentage width for the three panes
Previously the left and right panes on the bottom had fixed widths
totalling eight-hundred twenty pixels, while the center pane would
grow or shrink to take up remaining available space.  This layout
creates a problem when the width of the application is not much
greater than eight-hundred twenty; the center pane is so constricted
that it becomes difficult to use.

This patch sets the three panes to equal widths of thirty-three
percent, as an experiment to see if it produces better results.  It
makes the program more usable on my smaller screen, especially when
not maximized.  But it needs to be tested more on a variety of
screens.