Elide long stylesheet names by croppnig them in the middle
Very long sheet names look bad in the lower-right pane.  Here we crop
long ones by cutting out the middle part if needed, based on the
assumption that the most recognizable parts of the name come at either
the beginning or end, or both.  This change only affects the display
of the sheet names; it does not change the filename we use when saving
them.

Redmine-Ticket: 1760