Set the max-width and overfliew for the file pane
The maximum width ensures that the cropping effect on the filenames
will come into play.  Even so, we still also need to disable
overflow so that scroll bars will not appear.  That fact that we crop
the filename does not change the width that XUL thinks the label has,
and without the overflow setting we end up with long scrollbars for
labels that fit comfortably on screen to begin with.

Redmine-Ticket: 1760