Ignore local cache when browsing to pages
If we make changes to an existing page, browse away, and then come
back, the cached version can cause problems with our palette.  That is
because the palette is recalculated regardless of the cache, so
swatches with the same colors will be collapsed, which is not the
behavior we want.  We can use a flag from the nsIWebNavigation
interface to make sure we always grab a fresh copy of the page.

Redmine-Ticket: 1758