Reload pages when clicking ‘Forward’ or ‘Back’
A previous patch modified the way we load pages to avoid cached data,
but using the Forward and Back buttons still navigate through the
cache.  Therefore, we still run into problems with the palette if we
use those buttons.  This patch introduces a boolean into the Browser
that records when the user clicks either one.  Then inside of our
progress listener we forcefully reload the page if that boolean is
true, which ensures that we are not using cached pages.

Redmine-Ticket: 1758