Stop Array.top() from mixing up boolean false types
The existing implementation of Array.top() has two problems:

    1. If the array is empty it returns boolean false, whereas
    accessing out-of-bound elements in Javascript returns undefined.

    2. It returns false if the array has elements but if the top
    element is zero, ‘false’, or an empty string.

One way to fix this is to explicitly test whether or not the top
element is undefined, and if so then return undefined.  For example:

    Array.prototype.top = function () {
        var top = this.length - 1;
        return ( this[top] === undefined )
            ? undefined
            : this[top];
    };

However, we can take advantage of the previously stated behavior by
which Javascript returns undefined if we try to access an element that
does not exist.  We can therefore return whatever is on the top of the
array without having to perform any sanity checking.  This stops us
from doing things like converting zeros to boolean falses, and gives
us the benefit of returning undefined when the array is empty.