Cut short the loop in Color.getColorName()
Even after finding a color name for the given hex value, the for-loop
in getColorName() would continue to run through all of the names.
This patch makes it return as soon as it finds one.

The comment also claimed that it returned the hex if it could not find
a name, but the code itself did not do this.  It returned a boolean
false value if there was no match.  This commit updates the comment to
reflect the code logic, but at the same time also changes it to return
undefined instead of a boolean, as that makes more sense semantically
for a missing value.