ColorProperty: Redefine functions on the prototype for performance
Since it is costly to define these functions every time we instantiate
ColorProperty objects, we take out the functions and define them on
the prototype instead, so that we are not re-creating them every time.
Some of the functions originally referred to 'propertyType' and were
changed to refer to 'this.propertyType', since we no longer close over
the value passed into the constructor.