Palette: Rename insertStyle() to insertRule()
It is not obvious from the code or comments what kind of argument
insertStyle() accepts.  By looking at its one call point, inside of
csDocument.LoadPalette(), we see that it takes a CSSRule object.  This
patch clarifies that in the documentation and renames it to be more a
little more clear.