ColorProperty: Pass CSSStyleDeclaration object to constructor
The documentation claims that this is what the class accepts during
instantiation, but this has not been true since af8f59f9 (‘Refactor
interface for stylesheets’).  That commit introduced a bug by passing
a CSSRule to Palette.insertStyle() instead of the CSSStyleDeclaration
of that rule.

After that—unfortunately—the development of ColorProperty continued in
such a way as to support the use of the CSSRule object which it was
handed during instantiation.  But the whole time the class has been
documented and thought to be using a CSSStyleDeclaration.

This patch rectifies that error.  The insertStyle() function—now
called insertRule()—correctly passes the CSSStyleDeclaration to new
ColorProperty classes.  The setColor() and getColorString() methods
were written to operate on CSSRule objects.  So in this patch they are
rewritten to work on the CSSStyleDeclaration.