Fix the list of property names inside Palette.insertRule()
Now that we use getPropertyValue() to test whether or not a property
has a color, we need to use a different name format.  While the old
property() method expected names like ‘backgroundColor’, the
getPropertyValue() expects ‘background-color’.  This patch swaps out
the names for that format, so that we correctly load all of the color
properties.