Provide access to properties in a swatch
A swatch keeps track of all of the color properties that are affected
by a particular color.  All of those color property objects provide
access to the CSS rule object for which they are defined, which in
turn provide access to the CSS style sheet object in which the
property appears.  This means that color property objects allow us to
bridge the gap between a single color and the stylesheets which it
affects.

To facilitate that logic, this patch adds an accessor to the Swatch
class that gives us access to the color properties associated with it.