Refactor code for creating save-file dialog
This patch replaces CSApplication.getSaveFile() with a more generic
method that can display file dialogs for the purposes of both saving
and opening files.  This generic functionality is intended to
facilitate the creation of an open-file dialog, since it and the save
dialog will share much of the same code.