CSApplication: Make use of removeAllChildren()
Two functions, updatePalettePane() and updateToolPane(), contain loops
that remove all of the child elements from the palette and tool pane,
respectively.  Elsewhere in the application we define a
removeAllChildren() method on all Element objects for this purpose.
So these loops can be eliminated and replaced with a call to that
method.