Swatch: Rename ‘setLiveColorObserver’ to ‘removeSetColorObsever’
All of the methods related to observers are either add*Observer() or
remove*Observer().  The only odd-one out is setLiveColorObserver(),
which is also not currently used anywhere.  Looking at the logic it is
clear that it is intended to remove set-color observers.  So we rename
the method to reflect that.