Teach swatches how to accept colors pasted from the clipboard
We extend the Swatch class with a method that attempts to read a color
string from the clipboard and sets the swatch to that color.  The
color assignment happens via Swatch.setColor(), which will take a
string input and try to derive a color from that using
Color.from_css().  So the current mechanism will work for any string
accepted by from_css().

This patch also adds the Ctrl-V key-binding to the palette control,
which will fire off the pasting code.

Redmine-Ticket: 1785