Trim color strings when parsing CSS color values
Inside of from_css() we try to create a Color object from a string
describing that color.  It is possible that the input string will have
superfluous white-space at the beginning or end, or both.  For
example, if a user is copying a color string from his text editor, he
may accidentally copy the newline, or some extra space on one side.
Because a pasted color string will ultimately reach from_css(), we
want to trim off that white-space if it exists.