Use the ‘accel’ key modifier for keyboard shortcuts
For many shortcuts we use the Control key as the modifier.  This works
well on Windows and Linux, for instance, where Ctrl-O is the typical
key-binding for opening files.  However, on Macs the Ctrl-O shortcut
will not work, since that platform uses a different modifier key.

The XUL platform lets us get around the difference in modifier keys by
using the ‘accel’ modifier, which defaults to the appropriate modifier
key for the host platform.  This patch changes our existing shortcuts
to use ‘accel’ so that Mac users can use standard shortcuts for their
platform.