Return filename when saving files
The method CSApplication.save() takes a filename as one of its
parameters, which serves as the default name that appears in the file
selection dialog.  However, users may choose to save the file under a
different name.  It can be useful to know when this happens, and what
name the user chose in that situation.  So we modify save() to return
the name of the file in which the content was saved.