Teach csStyleSheet objects how to save to file
Previously we would call csApp.save() directly on the stylesheet
content when clicking the ‘Save’ button on the file pane.  However,
this approach makes it difficult to take full advantage of information
in the stylesheet when saving, primarily because all of our logic has
to be in the ‘oncommand’ attribute of the save button.  Here we add a
method to stylesheets so that they can implement their own logic for
saving, and the button now calls that instead.

For the time being, the new save() method replicates the previous
behavior of calling csApp.save() with the pretty-printed stylesheet
context.  So there is no behavioral change here, just a small change
to pave the way.