Link DOM stylesheet objects to our custom stylesheet objects
Some of the functionality of the application requires the ability to
find a path from a DOM CSS object to our own csStyleSheet class.  For
example, if we want to lock stylesheets or track certain
modifications, we want to be able to link ColorProperty objects to
csStyleSheet objects.  There is not a good place in our current
application structure to hook these two classes together; by the time
we instantiate one, we do not have access to the other.

One way to address this is to leverage an object that is shared
between color properties and our stylesheets: the DOM stylesheet
object.  Specifically objects of the CSSStyleSheet class.  By adding a
pointer on that class which points to our own stylesheet class, any
object that has a link to the DOM stylesheet can access our custom
stylesheet object.

With that in mind, we add a tentatively-named ‘colorFishSheet’ to the
DOM stylesheet.