Download and install dmg2img
This is a small program that is used to convert mac's dmg files into
mountable img files (effectively ISOs). Will download the source and compile
the program. The zlib-dev and libbz2-dev packages are required for the
compilation to work. Maybe I'll get some apt stuff in here... ha ha...
doubtful. Currently there are no dependancies for this so it won't do anything
different by calling "make all"

Redmine-Ticket: 1230