Add a checkbox in the file pane for toggling sheet locking
Users need to be able to toggle the locked status of each stylesheet,
and for the time being, the easiest way to do this is to put a simple
checkbox next to each sheet.  To facilitate the code behind this
process, we add a one-line method to csStyleSheet that toggles the
locked status.  That way we can avoid writing logic out-right in the
‘oncommand’ attribute of the checkbox, and can instead just call a
method.

Redmine-Ticket: 1780